requirements.txt
absl-py==1.4.0
accelerate==0.19.0
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
avalanche-lib==0.3.1
beautifulsoup4==4.12.2
brotlipy==0.7.0
cachetools==5.3.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
datasets==2.12.0
diffusers==0.16.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.1
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
grpcio==1.54.2
huggingface-hub==0.14.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.3
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
oauthlib==3.2.2
pandas==2.0.1
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.23.1
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
pytorchcv==0.0.67
pytz==2023.3
PyYAML==6.0
pyzmq==19.0.2
quadprog==0.1.11
regex==2023.5.5
requests-oauthlib==1.3.1
responses==0.18.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.23.1
setproctitle==1.3.2
smmap==5.0.0
soupsieve==2.4.1
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
tzdata==2023.3
wandb==0.15.3
Werkzeug==2.3.4
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
torch-fidelity==0.3.0